requirements.txt
python-telegram-bot[ext]~=20.7
aiohttp~=3.9
websockets~=12.0
pydantic~=2.5
loguru~=0.7
python-dotenv~=1.0
pytest-asyncio~=0.23
pydantic-settings>=2.0
tinydb>=4.8